fixed compiler warning
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@5608 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ io/include/pcl/compression/impl/octree_pointcloud_compression.hpp
@@ -108,7 +108,7 @@ namespace pcl
       // initialize color encoding
       colorCoder_.initializeEncoding ();
       colorCoder_.setPointCount (static_cast<unsigned int> (cloud_arg->points.size ()));
-      colorCoder_.setVoxelCount (this->leafCount_);
+      colorCoder_.setVoxelCount (static_cast<unsigned int> (this->leafCount_));
 
       // initialize point encoding
       pointCoder_.initializeEncoding ();
